Explicitly instantiate MWWorld::Store
